support Python 3.11, move configs to submodule
