+++ tox.ini
@@ -1,12 +1,13 @@
 [tox]
 minversion = 3.9.0
-envlist = py39, py310, flake8, mypy
+envlist = py39, py310, py311, flake8, mypy
 isolated_build = true
 
 [gh-actions]
 python =
     3.9: py39
     3.10: py310, mypy, flake8
+    3.11: py311
 
 [testenv]
 setenv =
